Allow a single published artifact to work with multiple Hadoop versions
This commit allows us to publish one `spark-avro` artifact which is built against a fixed Hadoop version but which works with both Hadoop 1.x and 2.x. In the past, we published separate artifacts for Hadoop 1.x and 2.x in order to work around a binary incompatibility in `TaskAttemptContext` (see #19). This patch works around the incompatibility using reflection, similar to https://github.com/apache/spark/pull/6599.

In order to make this testable, this patch also modifies our SBT build and Travis configuration so that the test Hadoop dependencies can be configured separately from the compile dependency.

I made a similar fix in `spark-redshift`: https://github.com/databricks/spark-redshift/pull/58

Author: Josh Rosen <joshrosen@databricks.com>

Closes #79 from JoshRosen/multiple-hadoop-version-support.