Ensure that FsInput is always closed in order to prevent resource leaks
There are two places where `FsInput`s are opened but never closed, leading to connection / resource leaks and causing hangs (such as the one reported in #156).

This patch addresses this issue by adding logic to ensure that these inputs are closed under all circumstances, including error-cases.

I managed to reproduce #156 in `spark-shell` and have verified that this fixes the hang.

Fixes #156.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #174 from JoshRosen/fix-leaked-s3-connections.