Fix ClassCastException when deserializing Map field
My Avro object has a field defined as java.util.Map.  In that case, item ends up being an instance of MapWrapper, which cannot be cast to a HashMap.  Casting to Map works fine, and presumably would still work in any case that you could cast to a HashMap.  I see nothing here that would specifically require the HashMap implementation.

Author: Todd Gibson <tgibson11@gmail.com>

Closes #190 from tgibson11/patch-1.