Add support for complex union types
A complex union will be converted to a struct type where field names are `member0`, `member1`, etc., in accordance with members of the union. This is consistent with the behavior when reading Parquet files. Field values are resolved following existing schema translation rules. The test case is a good example of this translation.

Author: Hao Xia <hao@optimizely.com>

Closes #117 from jasonxh/hao/complex-union.