Cast to `Iterable` instead of `GenericData.Array` in SchemaConverters
In the `ARRAY`-type handling code in `SchemaConverters` we always casted an array field's object to `GenericData.Array[AnyRef]`, which works fine as long as the Avro objects come from files but may break when they come in via other sources. To fix this, we can simply cast to `Iterable[AnyRef]` instead, since the code which uses the casted object is only calling methods from the `Iterable` interface.

This PR subsumes / closes #130 by defg.

Author: Josh Rosen <joshrosen@databricks.com>

Closes #194 from JoshRosen/array-type-handling.