Fix bug on generation of avro files with nested records with the same name
spark-avro was failing on the generation of avro files when nested records had the same name, as can be seen on the following issue [#54](https://github.com/databricks/spark-avro/issues/54).
For solving that problem has been added a different namespace on each record. This namespace is created with the concatenation of all the parents of each record separated by dot.
Added also a new test in AvroSuite for checking the new code.
This is related with pull request [#73](https://github.com/databricks/spark-avro/pull/73) but have been added a modification on AvroOutputWriter to solve correctly the issue.

Author: gsolasab <gerard.sola.sabate@everis.com>
Author: Gerard Solà <gerardsola@gmail.com>

Closes #249 from gsolasab/master.