Upgrade SBT and corresponding plugins. (#263)
This PR upgrades SBT to 0.13.16. It could not be upgraded to the latest version (1.0.4 as of now) because of [this issue](https://github.com/databricks/sbt-spark-package/issues/35) with `sbt-spark-package`. `sbt-idea` was removed from the plugins because, firstly, it is very old and would not support SBT 1+ and also because IntelliJ now [supports SBT](https://blog.jetbrains.com/scala/2017/03/23/scala-plugin-for-intellij-idea-2017-1-cleaner-ui-sbt-shell-repl-worksheet-akka-support-and-more/) which should be the correct way.